process-txt/requirements.txt
google-cloud-storage==2.10.0
google-cloud-bigquery==3.10.0